requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
gunicorn==21.2.0
idna==3.6
lxml==5.1.0
packaging==24.0
phonenumbers==8.13.32
pillow==10.2.0
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.0
typing-extensions==4.15.0
urllib3==2.5.0
